+++ sn_networking/src/lib.rs
@@ -321,6 +321,13 @@ impl Network {
                 return Ok(());
             }
             warn!("The obtained {n_verified} verified proofs did not match the expected {expected_n_verified} verified proofs");
+            // Sleep to avoid firing queries too close to even choke the nodes further.
+            let waiting_time = if retry_attempts == 1 {
+                MIN_WAIT_BEFORE_READING_A_PUT
+            } else {
+                MIN_WAIT_BEFORE_READING_A_PUT + MIN_WAIT_BEFORE_READING_A_PUT
+            };
+            tokio::time::sleep(waiting_time).await;
         }
 
         Err(Error::FailedToVerifyChunkProof(chunk_address.clone()))
@@ -410,58 +417,69 @@ impl Network {
         key: RecordKey,
         cfg: &GetRecordCfg,
     ) -> Result<Record> {
-        backoff::future::retry(ExponentialBackoff{
-            max_elapsed_time: Some(MAX_GET_RETRY_DURATION),
-            ..Default::default()
-        }, || async {
-            let pretty_key = PrettyPrintRecordKey::from(&key);
-            info!(
-                "Getting record of {pretty_key:?}. with cfg {cfg:?}",
-            );
-            let (sender, receiver) = oneshot::channel();
-            self.send_swarm_cmd(SwarmCmd::GetNetworkRecord {
-                key: key.clone(),
-                sender,
-                cfg: cfg.clone(),
-            }).map_err(|err| BackoffError::Transient { err,  retry_after: None })?;
-            let result = receiver.await.map_err(|e| {
+        backoff::future::retry(
+            ExponentialBackoff {
+                max_elapsed_time: Some(MAX_GET_RETRY_DURATION),
+                ..Default::default()
+            },
+            || async {
+                let pretty_key = PrettyPrintRecordKey::from(&key);
+                info!("Getting record from network of {pretty_key:?}. with cfg {cfg:?}",);
+                let (sender, receiver) = oneshot::channel();
+                self.send_swarm_cmd(SwarmCmd::GetNetworkRecord {
+                    key: key.clone(),
+                    sender,
+                    cfg: cfg.clone(),
+                })
+                .map_err(|err| BackoffError::Transient {
+                    err,
+                    retry_after: None,
+                })?;
+                let result = receiver.await.map_err(|e| {
                 error!("When fetching record {pretty_key:?}, encountered a channel error {e:?}");
                 Error::InternalMsgChannelDropped
             }).map_err(|err| BackoffError::Transient { err,  retry_after: None })?;
 
-            // log the results
-            match &result {
-                Ok(_) => {
-                    info!("Record returned: {pretty_key:?}. Retrying via backoff...");
-                }
-                Err(GetRecordError::RecordDoesNotMatch(_)) => {
-                    warn!("The returned record does not match target {pretty_key:?}. Retrying via backoff...");
-                }
-                Err(GetRecordError::NotEnoughCopies { expected, got, .. }) => {
-                    warn!("Not enough copies ({got}/{expected}) found yet for {pretty_key:?}. Retying via backoff...");
-                }
-                // libp2p RecordNotFound does mean no holders answered.
-                // it does not actually mean the record does not exist.
-                // just that those asked did not have it
-                Err(GetRecordError::RecordNotFound) => {
-                    warn!("No holder of record '{pretty_key:?}' found. Retrying via backoff...");
-                }
-                Err(GetRecordError::SplitRecord { .. }) => {
-                    error!("Encountered a split record for {pretty_key:?} Retrying via backoff...");
-                }
-                Err(GetRecordError::QueryTimeout) => {
-                    error!("Encountered query timeout for {pretty_key:?} Retrying via backoff...");
-                }
-            };
+                // log the results
+                match &result {
+                    Ok(_) => {
+                        info!("Record returned: {pretty_key:?}.");
+                    }
+                    Err(GetRecordError::RecordDoesNotMatch(_)) => {
+                        warn!("The returned record does not match target {pretty_key:?}.");
+                    }
+                    Err(GetRecordError::NotEnoughCopies { expected, got, .. }) => {
+                        warn!("Not enough copies ({got}/{expected}) found yet for {pretty_key:?}.");
+                    }
+                    // libp2p RecordNotFound does mean no holders answered.
+                    // it does not actually mean the record does not exist.
+                    // just that those asked did not have it
+                    Err(GetRecordError::RecordNotFound) => {
+                        warn!("No holder of record '{pretty_key:?}' found.");
+                    }
+                    Err(GetRecordError::SplitRecord { .. }) => {
+                        error!("Encountered a split record for {pretty_key:?}.");
+                    }
+                    Err(GetRecordError::QueryTimeout) => {
+                        error!("Encountered query timeout for {pretty_key:?}.");
+                    }
+                };
 
-            // if we dont want to retry, throw permanent error
-            if !cfg.re_attempt {
-                if let Err(e) = result {
-                    return Err(BackoffError::Permanent(Error::from(e)))
+                // if we dont want to retry, throw permanent error
+                if !cfg.re_attempt {
+                    if let Err(e) = result {
+                        return Err(BackoffError::Permanent(Error::from(e)));
+                    }
                 }
-            }
-            result.map_err(|err| BackoffError::Transient{err: Error::from(err), retry_after: None})
-        })
+                if result.is_err() {
+                    trace!("Getting record from network of {pretty_key:?} via backoff...");
+                }
+                result.map_err(|err| BackoffError::Transient {
+                    err: Error::from(err),
+                    retry_after: None,
+                })
+            },
+        )
         .await
     }
 
@@ -574,6 +592,8 @@ impl Network {
                 )
                 .await?;
             } else {
+                // The `verify_chunk_existence` will carry out three times of re-attempts
+                // However, `get_record_from_network` will only carry out once.
                 self.get_record_from_network(record.key.clone(), get_cfg)
                     .await?;
             }
@@ -633,16 +653,6 @@ impl Network {
             .map_err(|_e| Error::InternalMsgChannelDropped)
     }
 
-    // Add a list of keys of a holder to Replication Fetcher.
-    #[allow(clippy::mutable_key_type)] // for Bytes in NetworkAddress
-    pub fn add_keys_to_replication_fetcher(
-        &self,
-        holder: PeerId,
-        keys: HashMap<NetworkAddress, RecordType>,
-    ) -> Result<()> {
-        self.send_swarm_cmd(SwarmCmd::AddKeysToReplicationFetcher { holder, keys })
-    }
-
     /// Send `Request` to the given `PeerId` and await for the response. If `self` is the recipient,
     /// then the `Request` is forwarded to itself and handled, and a corresponding `Response` is created
     /// and returned to itself. Hence the flow remains the same and there is no branching at the upper
@@ -685,6 +695,10 @@ impl Network {
         self.send_swarm_cmd(SwarmCmd::GossipHandler)
     }
 
+    pub fn trigger_interval_replication(&self) -> Result<()> {
+        self.send_swarm_cmd(SwarmCmd::TriggerIntervalReplication)
+    }
+
     // Helper to send SwarmCmd
     fn send_swarm_cmd(&self, cmd: SwarmCmd) -> Result<()> {
         let capacity = self.swarm_cmd_sender.capacity();
@@ -692,20 +706,10 @@ impl Network {
         let cmd_sender = self.swarm_cmd_sender.clone();
 
         if capacity == 0 {
-            if matches!(cmd, SwarmCmd::AddKeysToReplicationFetcher { .. }) {
-                // we can safely drop AddKeysToReplicationFetcher
-                // it should be reattempted in a few seconds and if we can cope we'll do it.
-                warn!(
-                    "SwarmCmd channel is full. Dropping AddKeysToReplicationFetcher: {:?}",
-                    cmd
-                );
-                return Ok(());
-            } else {
-                error!(
-                    "SwarmCmd channel is full. Await capacity to send: {:?}",
-                    cmd
-                );
-            }
+            error!(
+                "SwarmCmd channel is full. Await capacity to send: {:?}",
+                cmd
+            );
         }
 
         // Spawn a task to send the SwarmCmd and keep this fn sync