+++ sn_testnet/src/lib.rs
@@ -228,6 +228,7 @@ impl Testnet {
 
         let rpc_address = "127.0.0.1:12001".parse()?;
         let mut launch_args = self.get_launch_args("safenode-1", Some(rpc_address), node_args)?;
+        launch_args.push("--first".to_string());
 
         let genesis_port: u16 = 11101;
         launch_args.push("--port".to_string());
@@ -438,6 +439,7 @@ mod test {
                     rpc_address.to_string(),
                     "--log-format".to_string(),
                     "json".to_string(),
+                    "--first".to_string(),
                     "--port".to_string(),
                     "11101".to_string(),
                 ]),
@@ -511,6 +513,7 @@ mod test {
                     rpc_address.to_string(),
                     "--log-format".to_string(),
                     "json".to_string(),
+                    "--first".to_string(),
                     "--port".to_string(),
                     "11101".to_string(),
                 ]),