+++ sn_testnet/CHANGELOG.md
@@ -2442,6 +2442,41 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Other
 - update dependencies
 
+## [0.2.383](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.2.382...sn_testnet-v0.2.383) - 2024-01-08
+
+### Other
+- update dependencies
+
+## [0.3.0](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.2.383...sn_testnet-v0.3.0) - 2024-01-08
+
+### Added
+- provide `--first` argument for `safenode`
+
+## [0.3.1](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.3.0...sn_testnet-v0.3.1) - 2024-01-09
+
+### Other
+- update dependencies
+
+## [0.3.2](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.3.1...sn_testnet-v0.3.2) - 2024-01-09
+
+### Other
+- update dependencies
+
+## [0.3.3](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.3.2...sn_testnet-v0.3.3) - 2024-01-09
+
+### Other
+- update dependencies
+
+## [0.3.4](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.3.3...sn_testnet-v0.3.4) - 2024-01-09
+
+### Other
+- update dependencies
+
+## [0.3.5](https://github.com/maidsafe/safe_network/compare/sn_testnet-v0.3.4...sn_testnet-v0.3.5) - 2024-01-10
+
+### Other
+- update dependencies
+
 ## v0.1.0 (2023-03-16)
 
 <csr-id-4f04bd1a5d1c747bfc6b5d39824dd108f8546b7b/>